pyproject.toml
[tool.poetry]
name = "mmpose"
version = "0.0.1"
description = ""
readme = "README.md"
authors = ["mountain-gorilla-inc"]

[tool.poetry.dependencies]
python = ">=3.9.2,<3.10"
invoke = "^1.7.0"
torch = {url = "https://download.pytorch.org/whl/cu111/torch-1.10.0%2Bcu111-cp39-cp39-linux_x86_64.whl"}
torchvision = {url = "https://download.pytorch.org/whl/cu111/torchvision-0.11.0%2Bcu111-cp39-cp39-linux_x86_64.whl"}
mmcv-full = {url = "https://download.openmmlab.com/mmcv/dist/cu111/torch1.10.0/mmcv_full-1.4.6-cp39-cp39-manylinux1_x86_64.whl"}
mmdet = "^2.28.1"
numpy = "^1.24.2"
matplotlib = "^3.7.0"
opencv-python = "^4.7.0.68"
scipy = "^1.10.1"
xtcocotools = "^1.13"
json-tricks = "^3.16.1"
munkres = "^1.1.4"

[tool.poetry.dev-dependencies]
mypy = "^0.950"
flake8 = "^4.0.1"
black = "^22.3.0"
isort = "^5.10.1"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
requests = "^2.27.1"

[tool.mypy]
python_version = 3.9
plugins = "pydantic.mypy"
check_untyped_defs = true
disallow_untyped_defs = true
ignore_errors = false
ignore_missing_imports = true
strict_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
